+++ tests/unit/conftest.py
@@ -15,7 +15,13 @@ from sqlglot.errors import SqlglotError, ParseError
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.errors import NotFound
 
-from databricks.labs.lakebridge.config import TranspileConfig
+from databricks.labs.lakebridge.config import (
+    TranspileConfig,
+    ReconcileConfig,
+    SourceConnectionConfig,
+    TargetConnectionConfig,
+    ReconcileMetadataConfig,
+)
 from databricks.labs.lakebridge.helpers.file_utils import make_dir
 from databricks.labs.lakebridge.transpiler.lsp.lsp_engine import LSPEngine
 from databricks.labs.lakebridge.transpiler.sqlglot.dialect_utils import SQLGLOT_DIALECTS
@@ -433,3 +439,47 @@ async def lsp_engine(test_resources: Path) -> AsyncGenerator[LSPEngine, None]:
     yield engine
     if engine.is_alive:
         await engine.shutdown()
+
+
+@pytest.fixture
+def oracle_recon_config() -> ReconcileConfig:
+    return ReconcileConfig(
+        report_type="all",
+        source=SourceConnectionConfig(
+            dialect="oracle",
+            catalog="ORCL",
+            schema="tpch_sf10009",
+            uc_connection_name="remorph_oracle9",
+        ),
+        target=TargetConnectionConfig(
+            catalog="tpch9",
+            schema="1000gb9",
+        ),
+        metadata_config=ReconcileMetadataConfig(
+            catalog="remorph9",
+            schema="reconcile9",
+            volume="reconcile_volume9",
+        ),
+    )
+
+
+@pytest.fixture
+def snowflake_recon_config() -> ReconcileConfig:
+    return ReconcileConfig(
+        report_type="all",
+        source=SourceConnectionConfig(
+            dialect="snowflake",
+            catalog="snowflake_sample_data9",
+            schema="tpch_sf10009",
+            uc_connection_name="remorph_snowflake9",
+        ),
+        target=TargetConnectionConfig(
+            catalog="tpch9",
+            schema="1000gb9",
+        ),
+        metadata_config=ReconcileMetadataConfig(
+            catalog="remorph9",
+            schema="reconcile9",
+            volume="reconcile_volume9",
+        ),
+    )
